fix: audio duration integer division, GPU interlaced crash
- audio_reader: duration was computed with integer division
  (totalSamplesPerChannel / originalSampleRate), truncating the fraction;
  e.g. 394176 samples @ 44100 Hz reported 8.0 s instead of 8.94 s, which
  corrupted every duration-derived value (effective sample rate, padding)
- cuda_mapped_frame: interlaced content hit LOG(FATAL) before mapping,
  making the decoder's Adaptive deinterlace mode dead code. Pass
  progressive_frame through instead; cuvidMapVideoFrame deinterlaces
  interlaced frames according to the decoder's DeinterlaceMode

+++ src/audio/audio_reader.cc
@@ -27,8 +27,9 @@ namespace decord {
             return;
         }
         avformat_close_input(&pFormatContext);
-        // Calculate accurate duration
-        duration = totalSamplesPerChannel / originalSampleRate;
+        // Calculate accurate duration (double division: totalSamplesPerChannel
+        // is an int, and integer division would truncate the fraction)
+        duration = static_cast<double>(totalSamplesPerChannel) / originalSampleRate;
         // Construct NDArray
         ToNDArray();
     }
